requirements.txt
gmpy2
pycryptodome
requests
libnum